chore: refactor MobileNav (AP-194)
Break up into smaller components, use labelMap and iconMap, and create a dir for subcomponents and index file for clean imports
