+++ src/app/components/MobileNav/index.ts
@@ -1,0 +1,3 @@
+export { default } from './MobileNav';
+export { default as Welcome } from './Welcome';
+export { default as DropdownLink } from './DropdownLink';
